UPGRADE : Laravel 13 - PHP 8.4
Vue 2 prevents upgrading dependencies

+++ config/two-factor.php
@@ -67,7 +67,7 @@ return [
 
     'confirm' => [
         'key' => '_2fa',
-        'time' => env('SESSION_LIFETIME', 60 * 3),
+        'time' => (int)env('SESSION_LIFETIME', 60 * 3),
     ],
 
     /*
